requirements-dev.txt
-r docs/requirements-docs.txt

pytest-cov
pytest-pep8
pytest-logging
tox
